Splitting the App into different widgets.
